Move StorageStats
